Test for reading file and sending this to the model
